requirements.txt
streamlit==1.32.0
Pillow==10.2.0 